+++ R/eng-app_server.R
@@ -41,4 +41,12 @@ eng_app_server <- function(input, output, session) {
     eng_mod_focus_20200318_piemonte_server("20200318_piemonte")
     eng_mod_focus_20200318_veneto_intensive_server("21")
     eng_mod_focus_20200320_novara_server("da_novara")
+    eng_mod_focus_20200320_novara_server("da_vercelli",
+                                     loc = "Vercelli",
+                                     pop = 174904
+    )
+    eng_mod_focus_20200320_novara_server("da_alessandria",
+                                     loc = "Alessandria",
+                                     pop = 428826
+    )
 }